requirements.txt
amqp==5.2.0
annotated-types==0.5.0
anyio==4.4.0
asgiref==3.6.0
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.2
billiard==4.2.0
black==23.11.0
Brotli==1.1.0
celery==5.4.0
certifi==2023.5.7
cffi==1.15.1
channels==4.1.0
channels-auth-token-middlewares==1.1.0
channels-redis==4.2.0
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.7
cssselect2==0.7.0
cssutils==2.9.0
daphne==4.1.2
Django==4.2.1
django-auditlog==2.3.0
django-cors-headers==4.2.0
django-filter==23.2
django-inlinecss @ git+https://github.com/blocher/django-inlinecss.git@73f6757a436c8b28487c9bef840cebf11ea33c45
django-jazzmin==2.6.0
django-weasyprint==2.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
drf-writable-nested==0.7.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
fonttools==4.42.1
future==0.18.3
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
httptools==0.6.1
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
JPype1==1.5.0
kombu==5.3.7
MarkupSafe==2.1.2
msgpack==1.0.8
mypy-extensions==1.0.0
nepali-datetime==1.0.7
num2words==0.5.13
numpy==1.26.4
openapi-codec==1.3.2
openpyxl==3.1.2
packaging==23.1
pandas==2.2.1
pathspec==0.12.1
Pillow==10.0.1
platformdirs==4.1.0
priority==1.3.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.7
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==2.4.1
pydantic_core==2.10.1
pydyf==0.8.0
PyJWT==2.7.0
pynliner==0.8.0
pyOpenSSL==24.1.0
pyphen==0.14.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.4
requests==2.30.0
service-identity==24.1.0
simplejson==3.19.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.4.4
tinycss2==1.2.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
uvicorn==0.23.2
vine==5.1.0
watchfiles==0.22.0
wcwidth==0.2.13
weasyprint==60.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
zope.interface==6.3
zopfli==0.2.3
